requirements.txt
sockets
pycryptodome
opencv-python